semi-functional location track
might have to move it back to home.dart so that it is accessible across pages
