+++ pubspec.yaml
@@ -31,6 +31,8 @@ dependencies:
   flutter:
     sdk: flutter
 
+  shared_preferences: ^2.3.2
+
   geolocator: ^13.0.1
   google_maps_flutter: ^2.9.0
   flutter_excel: ^1.0.1
